Ajout des tests pour le node
