+++ backend/babel.config.json
@@ -1,0 +1,3 @@
+{
+    "presets": ["@babel/preset-env"]
+}
