+++ backend/jest.config.js
@@ -1,0 +1,7 @@
+export default {
+    transform: {
+      "^.+\\.js$": "babel-jest"
+    },
+    testEnvironment: "node",
+  };
+
